Undo flip transform and debug magnetometer
Turns out the flip transform can already be done with axis remaps. This solves some other problems I've been having as well.

I'm still struggling to get magnetometer orientation correct though.

+++ Firmware/SlimeVR-Tracker-ESP-0.5.2/src/sensors/SensorManager.h
@@ -110,7 +110,8 @@ private:
 			return sensor;
 		}
 
-		uint8_t intPin = extraParam;
+		//bug: bad cast from int to u8: our remap config never makes it in here because of this.
+		int intPin = extraParam;
 		sensor = std::make_unique<ImuType>(
 			sensorID,
 			addrSuppl,
